refactor : 메소드명 수정 및 기타 피드백 수정사항 반영
 - dto 패키지 변경
 - isHittable로 메소드명 변경
 - getCards를 통해 바로 List 반환
 - hit 메소드를 통해 Deck을 바로 인자로 받도록 수정
 - 접근 제어자 수정
